Ponderada4/teste/requirements.txt
pandas
matplotlib
psycopg2-binary
fastapi
uvicorn
jinja2